requirements.txt
fastapi==0.115.11
uvicorn==0.33.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.1.0
httpx==0.27.0
python-json-logger==3.3.0
pytest==7.3.1
pytest-cov==4.1.0